requirements.txt
aiosmtpd==1.4.2
atpublic==3.0.1
attrs==21.4.0
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
dash==2.4.1
dash-bootstrap-components==1.1.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-labs==1.0.8
dash-table==5.0.0
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Compress==1.12
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
libretranslatepy==2.1.1
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.29
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
plotly==5.8.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
tenacity==8.0.1
translate==3.6.1
urllib3==1.26.9
Werkzeug==2.1.2
